requirements.txt
adabound==0.0.5
-e git+git@github.com:sh8/rdopt.git@7601bca4818a03ef1ace1e0c1df396ccec56003f#egg=camera_jacobian&subdirectory=lib/csrc/camera_jacobian
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cycler==0.10.0
Cython==0.29.17
dataclasses==0.8
decorator==4.4.2
docker-pycreds==0.4.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
googledrivedownloader==0.4
gql==0.2.0
graphql-core==1.1
greenlet==0.4.17
h5py==2.10.0
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
imageio-ffmpeg==0.4.2
imread==0.7.1
isodate==0.6.0
joblib==0.14.1
kiwisolver==1.2.0
kornia==0.2.0
llvmlite==0.32.0
matplotlib==3.2.2
moviepy==1.0.3
msgpack==1.0.0
networkx==2.4
neural-renderer-pytorch==1.1.3
numba==0.49.0
numpy==1.19.0
nvidia-ml-py3==7.352.0
opencv-python==4.2.0.34
openmesh==1.1.3
pandas==1.0.3
pathtools==0.1.2
Pillow==7.1.2
plyfile==0.7.2
proglog==0.1.9
promise==2.3
protobuf==3.13.0
psutil==5.7.0
pycocotools==2.0.0
pycparser==2.20
pynvim==0.4.2
PyOpenGL==3.1.5
pyparsing==3.0.0a1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
rdflib==5.0.0
requests==2.23.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.11.0
sentry-sdk==0.15.1
shortuuid==1.0.1
six==1.14.0
smmap==3.0.4
soft-renderer==1.0.0
subprocess32==3.5.4
tensorboardX==2.0
threadpoolctl==2.0.0
tifffile==2020.5.11
torch==1.9.0
torchvision==0.10.0
tqdm==4.46.0
transforms3d==0.3.1
typing-extensions==3.10.0.0
urllib3==1.25.9
wandb==0.10.15
watchdog==0.10.2
Werkzeug==1.0.1
wget==3.2
yacs==0.1.7
